pulled out instancing into separate file
