+++ src/lib.rs
@@ -6,72 +6,52 @@
 #![feature(async_closure)]
 #![feature(stmt_expr_attributes)]
 // #![feature(let_chains)]
-use crate::ui::UrlBar;
-use bevy::{
-	asset::load_internal_asset, ecs as bevy_ecs, prelude::*, reflect::TypeUuid,
-	render::primitives::Frustum,
-};
 #[cfg(target_arch = "wasm32")]
 use core::future::Future;
 // use core::slice::SlicePattern;
-use serde::{Deserialize, Serialize};
 // use bevy::winit::WinitSettings;
-use bevy_ecs::prelude::Component;
-use bevy_egui::EguiPlugin;
 #[cfg(feature = "normalmouse")]
 use bevy_flycam::{FlyCam, MovementSettings, NoCameraPlayerPlugin};
 //use bevy_kira_audio::AudioPlugin;
 // use bevy_inspector_egui::{Inspectable, InspectorPlugin};
-use bevy_mod_picking::*;
 //use bevy_egui::render_systems::ExtractedWindowSizes;
 //use bevy::window::PresentMode;
-use bevy::diagnostic::FrameTimeDiagnosticsPlugin;
 
 #[cfg(target_arch = "wasm32")]
 use gloo_worker::Spawnable;
 // use bevy::diagnostic::LogDiagnosticsPlugin;
-use crate::movement::Destination;
 #[cfg(feature = "adaptive-fps")]
 use bevy::diagnostic::Diagnostics;
-use bevy::{render::primitives::Sphere, window::RequestRedraw};
+
+use crate::{instancing::CustomMaterialPlugin, movement::Destination, ui::UrlBar};
+use bevy::{
+	asset::load_internal_asset,
+	diagnostic::FrameTimeDiagnosticsPlugin,
+	ecs as bevy_ecs,
+	prelude::*,
+	reflect::TypeUuid,
+	render::view::{ComputedVisibility, Msaa, NoFrustumCulling, Visibility},
+	window::RequestRedraw,
+};
+use bevy_ecs::prelude::Component;
+use bevy_egui::EguiPlugin;
+use bevy_mod_picking::*;
 use bevy_polyline::{prelude::*, PolylinePlugin};
-use std::f32::consts::PI;
-// use scale_info::build;
+use chrono::prelude::*;
+use datasource::DataUpdate;
+use primitive_types::H256;
+use serde::{Deserialize, Serialize};
 use std::{
 	collections::HashMap,
+	convert::AsRef,
+	f32::consts::PI,
 	num::NonZeroU32,
 	sync::{
-		atomic::{AtomicU32, Ordering},
-		Arc,
+		atomic::{AtomicI32, AtomicU32, Ordering},
+		Arc, Mutex,
 	},
+	time::Duration,
 };
-
-use bevy::{
-	core_pipeline::core_3d::Transparent3d,
-	ecs::system::{lifetimeless::*, SystemParamItem},
-	math::prelude::*,
-	pbr::{MeshPipeline, MeshPipelineKey, MeshUniform, SetMeshBindGroup, SetMeshViewBindGroup},
-	prelude::*,
-	render::{
-		extract_component::{ExtractComponent, ExtractComponentPlugin},
-		mesh::{GpuBufferInfo, MeshVertexBufferLayout},
-		render_asset::RenderAssets,
-		render_phase::{
-			AddRenderCommand, DrawFunctions, EntityRenderCommand, RenderCommandResult, RenderPhase,
-			SetItemPipeline, TrackedRenderPass,
-		},
-		render_resource::*,
-		renderer::RenderDevice,
-		view::{ComputedVisibility, ExtractedView, Msaa, NoFrustumCulling, Visibility},
-		RenderApp, RenderStage,
-	},
-};
-use bytemuck::{Pod, Zeroable};
-
-// use bevy_instancing::prelude::{
-//     ColorMeshInstance, CustomMaterial, CustomMaterialPlugin, IndirectRenderingPlugin,
-//     InstanceCompute, InstanceComputePlugin, InstanceSlice,
-// InstanceSliceBundle,BasicMaterialPlugin,TextureMaterialPlugin };
 
 #[cfg(feature = "atmosphere")]
 use bevy_atmosphere::prelude::*;
@@ -79,13 +59,14 @@ use bevy_atmosphere::prelude::*;
 // use bevy::diagnostic::FrameTimeDiagnosticsPlugin;
 // use ui::doturl;
 //use bevy_kira_audio::Audio;
-use std::{sync::Mutex, time::Duration};
 mod content;
-use std::sync::atomic::AtomicI32;
 mod datasource;
+mod instancing;
 mod movement;
 mod style;
 mod ui;
+
+use instancing::{InstanceData, InstanceMaterialData};
 
 #[cfg(target_family = "wasm")]
 pub mod webworker;
@@ -98,10 +79,6 @@ use crate::ui::{Details, DotUrl};
 // use bevy::winit::WinitSettings;
 #[cfg(feature = "spacemouse")]
 use bevy_spacemouse::{SpaceMousePlugin, SpaceMouseRelativeControllable};
-use chrono::prelude::*;
-use datasource::DataUpdate;
-use primitive_types::H256;
-use std::convert::AsRef;
 // #[subxt::subxt(runtime_metadata_path = "polkadot_metadata.scale")]
 // pub mod polkadot {}
 pub mod recorder;
@@ -1444,7 +1421,7 @@ fn add_blocks(
 					// 	value: block.details().value
 					// })
 					.insert(ClearMe)
-					.insert(Rainable { dest: base_y + target_y * build_dir, build_direction })
+					// .insert(Rainable { dest: base_y + target_y * build_dir, build_direction })
 					.insert(Name::new("Extrinsic"))
 					.insert(MedFi);
 
@@ -1580,204 +1557,6 @@ fn add_blocks(
 	}
 }
 
-#[derive(Component, Deref)]
-struct InstanceMaterialData(Vec<InstanceData>);
-
-impl ExtractComponent for InstanceMaterialData {
-	type Query = (&'static InstanceMaterialData, &'static Frustum);
-	type Filter = ();
-
-	// Frustum cull at extract
-	fn extract_component(item: bevy::ecs::query::QueryItem<Self::Query>) -> Self {
-		//Todo: make frustum planes slightly bigger by half a box width so sphere can be point.
-		InstanceMaterialData(
-			item.0
-				 .0
-				.iter()
-				.filter_map(|i| {
-					item.1
-						.intersects_sphere(&Sphere { center: i.position.into(), radius: 5. }, false)
-						.then(|| i.clone())
-				})
-				.collect::<Vec<_>>(),
-		)
-	}
-}
-
-pub struct CustomMaterialPlugin;
-
-impl Plugin for CustomMaterialPlugin {
-	fn build(&self, app: &mut App) {
-		app.add_plugin(ExtractComponentPlugin::<InstanceMaterialData>::default());
-		app.sub_app_mut(RenderApp)
-			.add_render_command::<Transparent3d, DrawCustom>()
-			.init_resource::<CustomPipeline>()
-			.init_resource::<SpecializedMeshPipelines<CustomPipeline>>()
-			.add_system_to_stage(RenderStage::Queue, queue_custom)
-			.add_system_to_stage(RenderStage::Prepare, prepare_instance_buffers);
-	}
-}
-
-#[derive(Clone, Copy, Pod, Zeroable)]
-#[repr(C)]
-struct InstanceData {
-	position: Vec3,
-	scale: f32,
-	color: [f32; 4],
-	// TODO: color can be a u32!
-	//  // Unpack the `u32` from the vertex buffer into the `vec4<f32>` used by the fragment shader
-	//    out.color = vec4<f32>((vec4<u32>(vertex.color) >> vec4<u32>(0u, 8u, 16u, 24u)) &
-	// vec4<u32>(255u)) / 255.0;
-}
-
-#[allow(clippy::too_many_arguments)]
-fn queue_custom(
-	transparent_3d_draw_functions: Res<DrawFunctions<Transparent3d>>,
-	custom_pipeline: Res<CustomPipeline>,
-	msaa: Res<Msaa>,
-	mut pipelines: ResMut<SpecializedMeshPipelines<CustomPipeline>>,
-	mut pipeline_cache: ResMut<PipelineCache>,
-	meshes: Res<RenderAssets<Mesh>>,
-	material_meshes: Query<(Entity, &MeshUniform, &Handle<Mesh>), With<InstanceMaterialData>>,
-	mut views: Query<(&ExtractedView, &mut RenderPhase<Transparent3d>)>,
-) {
-	let draw_custom = transparent_3d_draw_functions.read().get_id::<DrawCustom>().unwrap();
-
-	let msaa_key = MeshPipelineKey::from_msaa_samples(msaa.samples);
-
-	for (view, mut transparent_phase) in &mut views {
-		let rangefinder = view.rangefinder3d();
-		for (entity, mesh_uniform, mesh_handle) in &material_meshes {
-			if let Some(mesh) = meshes.get(mesh_handle) {
-				let key =
-					msaa_key | MeshPipelineKey::from_primitive_topology(mesh.primitive_topology);
-				let pipeline = pipelines
-					.specialize(&mut pipeline_cache, &custom_pipeline, key, &mesh.layout)
-					.unwrap();
-				transparent_phase.add(Transparent3d {
-					entity,
-					pipeline,
-					draw_function: draw_custom,
-					distance: rangefinder.distance(&mesh_uniform.transform),
-				});
-			}
-		}
-	}
-}
-
-#[derive(Component)]
-pub struct InstanceBuffer {
-	buffer: Buffer,
-	length: usize,
-}
-
-fn prepare_instance_buffers(
-	mut commands: Commands,
-	query: Query<(Entity, &InstanceMaterialData)>,
-	render_device: Res<RenderDevice>,
-) {
-	for (entity, instance_data) in &query {
-		let buffer = render_device.create_buffer_with_data(&BufferInitDescriptor {
-			label: Some("instance data buffer"),
-			contents: bytemuck::cast_slice(instance_data.as_slice()),
-			usage: BufferUsages::VERTEX | BufferUsages::COPY_DST,
-		});
-		commands
-			.entity(entity)
-			.insert(InstanceBuffer { buffer, length: instance_data.len() });
-	}
-}
-
-pub struct CustomPipeline {
-	shader: Handle<Shader>,
-	mesh_pipeline: MeshPipeline,
-}
-
-impl FromWorld for CustomPipeline {
-	fn from_world(world: &mut World) -> Self {
-		CustomPipeline {
-			shader: SHADER_HANDLE.typed(),
-			mesh_pipeline: world.resource::<MeshPipeline>().clone(),
-		}
-	}
-}
-
-impl SpecializedMeshPipeline for CustomPipeline {
-	type Key = MeshPipelineKey;
-
-	fn specialize(
-		&self,
-		key: Self::Key,
-		layout: &MeshVertexBufferLayout,
-	) -> Result<RenderPipelineDescriptor, SpecializedMeshPipelineError> {
-		let mut descriptor = self.mesh_pipeline.specialize(key, layout)?;
-		descriptor.vertex.shader = self.shader.clone();
-		descriptor.vertex.buffers.push(VertexBufferLayout {
-			array_stride: std::mem::size_of::<InstanceData>() as u64,
-			step_mode: VertexStepMode::Instance,
-			attributes: vec![
-				VertexAttribute {
-					format: VertexFormat::Float32x4,
-					offset: 0,
-					shader_location: 3, /* shader locations 0-2 are taken up by Position, Normal
-					                     * and UV attributes */
-				},
-				VertexAttribute {
-					format: VertexFormat::Float32x4,
-					offset: VertexFormat::Float32x4.size(),
-					shader_location: 4,
-				},
-			],
-		});
-		descriptor.fragment.as_mut().unwrap().shader = self.shader.clone();
-		descriptor.layout = Some(vec![
-			self.mesh_pipeline.view_layout.clone(),
-			self.mesh_pipeline.mesh_layout.clone(),
-		]);
-
-		Ok(descriptor)
-	}
-}
-
-type DrawCustom =
-	(SetItemPipeline, SetMeshViewBindGroup<0>, SetMeshBindGroup<1>, DrawMeshInstanced);
-
-pub struct DrawMeshInstanced;
-
-impl EntityRenderCommand for DrawMeshInstanced {
-	type Param =
-		(SRes<RenderAssets<Mesh>>, SQuery<Read<Handle<Mesh>>>, SQuery<Read<InstanceBuffer>>);
-	#[inline]
-	fn render<'w>(
-		_view: Entity,
-		item: Entity,
-		(meshes, mesh_query, instance_buffer_query): SystemParamItem<'w, '_, Self::Param>,
-		pass: &mut TrackedRenderPass<'w>,
-	) -> RenderCommandResult {
-		let mesh_handle = mesh_query.get(item).unwrap();
-		let instance_buffer = instance_buffer_query.get_inner(item).unwrap();
-
-		let gpu_mesh = match meshes.into_inner().get(mesh_handle) {
-			Some(gpu_mesh) => gpu_mesh,
-			None => return RenderCommandResult::Failure,
-		};
-
-		pass.set_vertex_buffer(0, gpu_mesh.vertex_buffer.slice(..));
-		pass.set_vertex_buffer(1, instance_buffer.buffer.slice(..));
-
-		match &gpu_mesh.buffer_info {
-			GpuBufferInfo::Indexed { buffer, index_format, count } => {
-				pass.set_index_buffer(buffer.slice(..), 0, *index_format);
-				pass.draw_indexed(0..*count, 0, 0..instance_buffer.length as u32);
-			},
-			GpuBufferInfo::NonIndexed { vertex_count } => {
-				pass.draw(0..*vertex_count, 0..instance_buffer.length as u32);
-			},
-		}
-		RenderCommandResult::Success
-	}
-}
-
 /// Yes this is now a verb. Who knew?
 fn rainbow(vertices: &mut Vec<Vec3>, points: usize) {
 	let start = vertices[0];
@@ -1807,11 +1586,11 @@ fn rainbow(vertices: &mut Vec<Vec3>, points: usize) {
 	}
 }
 
-#[derive(Component)]
-pub struct Rainable {
-	dest: f32,
-	build_direction: BuildDirection,
-}
+// #[derive(Component)]
+// pub struct Rainable {
+// 	dest: f32,
+// 	build_direction: BuildDirection,
+// }
 
 #[derive(Component, Deref, DerefMut)]
 struct AnimationTimer(Timer);
@@ -1842,12 +1621,9 @@ macro_rules! min {
 
 fn rain(
 	time: Res<Time>,
-	mut commands: Commands,
-	// mut drops: Query<(Entity, &mut Transform, &Rainable)>,
 	mut drops: Query<(Entity, &mut InstanceMaterialData)>,
 	mut timer: ResMut<UpdateTimer>,
 ) {
-	//TODO: remove the Rainable component once it has landed for performance!
 	let delta = 1.;
 	if timer.timer.tick(time.delta()).just_finished() {
 		for (entity, mut rainable) in drops.iter_mut() {
